moved things around to avoid possible inclusion of asynchronous coding
